Powered Rails now works very well
Fixed rolling amplitude
Riding position rewrite (Not tested yet)
Interaction will interact much better

+++ src/main/java/cn/nukkit/entity/item/EntityBoat.java
@@ -97,7 +97,7 @@ public class EntityBoat extends EntityVehicle {
                 return false;
             }
             // Event stop
-            performHurtAnimation((float) event.getDamage());
+            performHurtAnimation((int) event.getDamage());
 
             Entity damager = ((EntityDamageByEntityEvent) source).getDamager();
             boolean instantKill = damager instanceof Player
